Added EventSumController + Tests and some refactor
