+++ src/main/java/ar/com/tacsutn/grupo1/eventapp/web/EventSumController.java
@@ -1,0 +1,18 @@
+package ar.com.tacsutn.grupo1.eventapp.web;
+
+import ar.com.tacsutn.grupo1.eventapp.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class EventSumController implements Controller {
+    @GetMapping("/events/total")
+    public MockupResponse getV1(@RequestParam(value = "from", defaultValue = "01/01/1970") String fromDate,
+                                @RequestParam(value = "to", defaultValue = "31/12/9999") String toDate) {
+
+        // TODO: Add logic
+
+        return new MockupResponse("Se registraron 100 eventos desde la fecha " + fromDate + " hasta " + toDate);
+    }
+}
